wf_requirements.txt
svgpathtools==1.6.1
svgwrite==1.4.3
pyyaml==6.0.2